// generation/generated/typescript/runtime.ts
export const BASE_PATH = "/api/rest".replace(/\/+$/, "");

export interface ConfigurationParameters {
    basePath?: (string) | undefined; // override base path
    fetchApi?: (FetchAPI) | undefined; // override for fetch implementation
    middleware?: (Middleware[]) | undefined; // middleware to apply before/after fetch requests
    queryParamsStringify?: ((params: HTTPQuery) => string) | undefined; // stringify function for query strings
    username?: (string) | undefined; // parameter for basic security
    password?: (string) | undefined; // parameter for basic security
    apiKey?: (string | Promise<string> | ((name: string) => string | Promise<string>)) | undefined; // parameter for apiKey security
    accessToken?: (string | Promise<string> | ((name?: string, scopes?: string[]) => string | Promise<string>)) | undefined; // parameter for oauth2 security
    headers?: (HTTPHeaders) | undefined; //header params we want to use on every request
    credentials?: (RequestCredentials) | undefined; //value for the credentials param we want to use on each request
}

export class Configuration {
    constructor(private configuration: ConfigurationParameters = {}) {}

    set config(configuration: Configuration) {
        this.configuration = configuration;
    }

    get basePath(): string {
        return this.configuration.basePath != null ? this.configuration.basePath : BASE_PATH;
    }

    get fetchApi(): FetchAPI | undefined {
        return this.configuration.fetchApi;
    }

    get middleware(): Middleware[] {
        return this.configuration.middleware || [];
    }

    get queryParamsStringify(): (params: HTTPQuery) => string {
        return this.configuration.queryParamsStringify || querystring;
    }

    get username(): string | undefined {
        return this.configuration.username;
    }

    get password(): string | undefined {
        return this.configuration.password;
    }

    get apiKey(): ((name: string) => string | Promise<string>) | undefined {
        const apiKey = this.configuration.apiKey;
        if (apiKey) {
            return typeof apiKey === 'function' ? apiKey : () => apiKey;
        }
        return undefined;
    }

    get accessToken(): ((name?: string, scopes?: string[]) => string | Promise<string>) | undefined {
        const accessToken = this.configuration.accessToken;
        if (accessToken) {
            return typeof accessToken === 'function' ? accessToken : async () => accessToken;
        }
        return undefined;
    }

    get headers(): HTTPHeaders | undefined {
        return this.configuration.headers;
    }

    get credentials(): RequestCredentials | undefined {
        return this.configuration.credentials;
    }
}
// ...
export class FetchError extends Error {
    override name: "FetchError" = "FetchError";
    constructor(public cause: Error, msg?: string) {
        super(msg);

        // restore prototype chain
        const actualProto = new.target.prototype;
        if (Object.setPrototypeOf) {
            Object.setPrototypeOf(this, actualProto);
        }
    }
}
// ...
export type FetchAPI = WindowOrWorkerGlobalScope['fetch'];

export type Json = any;
export type HTTPMethod = 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE' | 'OPTIONS' | 'HEAD';
export type HTTPHeaders = { [key: string]: string };
export type HTTPQuery = { [key: string]: string | number | null | boolean | Array<string | number | null | boolean> | Set<string | number | null | boolean> | HTTPQuery };
// ...
export interface FetchParams {
    url: string;
    init: RequestInit;
}
// ...
export interface RequestContext {
    fetch: FetchAPI;
    url: string;
    init: RequestInit;
}

export interface ResponseContext {
    fetch: FetchAPI;
    url: string;
    init: RequestInit;
    response: Response;
}

export interface ErrorContext {
    fetch: FetchAPI;
    url: string;
    init: RequestInit;
    error: unknown;
    response?: (Response) | undefined;
}

export interface Middleware {
    pre?: ((context: RequestContext) => Promise<FetchParams | void>) | undefined;
    post?: ((context: ResponseContext) => Promise<Response | void>) | undefined;
    onError?: ((context: ErrorContext) => Promise<Response | void>) | undefined;
}